+++ kommunicate/src/main/java/io/kommunicate/Kommunicate.java
@@ -62,6 +62,7 @@ import io.kommunicate.database.KmDatabaseHelper;
 import io.kommunicate.models.KmAppSettingModel;
 import io.kommunicate.models.KmPrechatInputModel;
 import io.kommunicate.preference.KmPreference;
+import io.kommunicate.zendesk.KmZendeskClient;
 import io.kommunicate.users.KMUser;
 import io.kommunicate.utils.KmConstants;
 import io.kommunicate.utils.KmUtils;
@@ -827,6 +828,7 @@ public class Kommunicate {
                             }
                         }
                     }
+
                     @Override
                     public void onFailure(Object error) {
 
@@ -894,7 +896,14 @@ public class Kommunicate {
         new SecureSharedPreferences(AlPrefSettings.AL_PREF_SETTING_KEY, ApplozicService.getContext(context)).edit().remove("APPLICATION_KEY").commit();
     }
 
+
     public static void setChatText(Context context, String PreFilledText) {
         BroadcastService.onAutoText(context,PreFilledText);
+    }
+  
+    public static void openZendeskChat(final Context context) {
+        KmZendeskClient kmZendeskClient = KmZendeskClient.getInstance(context);
+        kmZendeskClient.openZendeskChat(context);
+
     }
 }