+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/conversation/fragment/MobiComConversationFragment.java
@@ -29,6 +29,7 @@ import android.text.Editable;
 import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.util.DisplayMetrics;
+import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.KeyEvent;
@@ -201,6 +202,8 @@ import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
 import de.hdodenhof.circleimageview.CircleImageView;
+import io.kommunicate.KmConversationBuilder;
+import io.kommunicate.KmConversationHelper;
 import io.kommunicate.preference.KmBotPreference;
 import io.kommunicate.KmSettings;
 import io.kommunicate.Kommunicate;
@@ -216,9 +219,12 @@ import io.kommunicate.callbacks.KmRemoveMemberCallback;
 import io.kommunicate.database.KmAutoSuggestionDatabase;
 import io.kommunicate.models.KmApiResponse;
 import io.kommunicate.models.KmFeedback;
+
 import io.kommunicate.preference.KmConversationInfoSetting;
+
 import io.kommunicate.services.KmClientService;
 import io.kommunicate.services.KmService;
+import io.kommunicate.zendesk.KmZendeskClient;
 import io.kommunicate.utils.KmAppSettingPreferences;
 import io.kommunicate.utils.KmInputTextLimitUtil;
 import io.kommunicate.utils.KmUtils;
@@ -379,6 +385,7 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
     protected boolean isUserGivingEmail;
     protected RelativeLayout conversationRootLayout;
     protected KmConversationInfoView kmConversationInfoView;
+    private String zendeskChatSdk;
     public static final int STANDARD_HEX_COLOR_CODE_LENGTH = 7;
     public static final int STANDARD_HEX_COLOR_CODE_WITH_OPACITY_LENGTH = 9;
     public boolean isCustomToolbarSubtitleDesign;
@@ -460,6 +467,9 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
         messageImageLoader.setImageFadeIn(false);
         messageImageLoader.addImageCache((getActivity()).getSupportFragmentManager(), 0.1f);
         kmAudioRecordManager = new KmAudioRecordManager(getActivity());
+
+        zendeskChatSdk = KmAppSettingPreferences.getInstance().getZendeskSdkKey();
+
     }
 
     private void setupChatBackground() {
@@ -577,7 +587,22 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
             @Override
             public void onRestartConversationPressed() {
                 AlEventManager.getInstance().sendOnConversationRestartedEvent(channel != null ? channel.getKey() : 0);
-                setFeedbackDisplay(false);
+                if(!TextUtils.isEmpty(zendeskChatSdk)) {
+                    KmConversationHelper.launchConversationIfLoggedIn(getContext(), new KmCallback() {
+                        @Override
+                        public void onSuccess(Object message) {
+                            setFeedbackDisplay(false);
+                            Utils.printLog(getContext(), TAG, "Successfully started a new Zendesk chat : " + (Integer) message);
+                        }
+
+                        @Override
+                        public void onFailure(Object error) {
+                            setFeedbackDisplay(false);
+                        }
+                    });
+                } else {
+                    setFeedbackDisplay(false);
+                }
             }
         });
 
@@ -1127,7 +1152,11 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
             return;
         }
 
+
         final String existingAssignee = channel.getConversationAssignee();
+
+        String existingAssignee = channel.getConversationAssignee();
+
         channel = ChannelService.getInstance(getActivity()).getChannelByChannelKey(channel.getKey());
         //conversation is open
         //if the conversation is opened from the dashboard while the feedback input fragment is open, the feedback fragment will be closed
@@ -1141,6 +1170,16 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
             }
         });
 
+
+        if (existingAssignee != null && !existingAssignee.equals(channel.getConversationAssignee())) {
+            showAwayMessage(true, null);
+        }
+
+        //If user has Integrated Zopim, initialize Zendesk Chat SDK
+        Contact assigneeContact = appContactService.getContactById(channel.getConversationAssignee());
+        if(!TextUtils.isEmpty(zendeskChatSdk) && assigneeContact != null && User.RoleType.AGENT.getValue().equals(assigneeContact.getRoleType()) && channel.getKmStatus() != Channel.CLOSED_CONVERSATIONS) {
+            KmZendeskClient.getInstance(getContext()).handleHandoff(channel, true);
+        }
     }
 
     @Override
@@ -1327,6 +1366,9 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
     }
 
     protected void processSendMessage() {
+        if(KmZendeskClient.getInstance(getContext()).isZendeskInitialized()) {
+            KmZendeskClient.getInstance(getContext()).sendZendeskMessage(messageEditText.getText().toString());
+        }
         if (!TextUtils.isEmpty(messageEditText.getText().toString().trim()) || !TextUtils.isEmpty(filePath)) {
             String inputMessage = messageEditText.getText().toString();
             String[] inputMsg = inputMessage.toLowerCase().split(" ");
@@ -2278,6 +2320,10 @@ public abstract class MobiComConversationFragment extends Fragment implements Vi
                 && messageTemplate != null && messageTemplate.isEnabled() && templateAdapter != null) {
             templateAdapter.setMessageList(new HashMap<String, String>());
             templateAdapter.notifyDataSetChanged();
+        }
+        if(!TextUtils.isEmpty(zendeskChatSdk)) {
+            KmZendeskClient.getInstance(getContext()).initializeZendesk(zendeskChatSdk, appContactService.getContactById(MobiComUserPreference.getInstance(getContext()).getUserId()));
+            MobiComUserPreference.getInstance(getContext()).setLatestZendeskConversationId(channel.getKey());
         }
     }